Modify Code

+++ Springboot-starterkit/src/main/java/com/busreservationsystem/application/repository/AgencyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.busreservationsystem.application.dto.common.AgencyDto;
 
 @Repository
-public interface AgencyRepository extends JpaRepository<AgencyDto ,Long>{
+public interface AgencyRepository extends JpaRepository<AgencyDto ,Integer>{
 
 
 }
